--- src/views/material/MaterialDepartement.vue
@@ -1,19 +1,0 @@
-<template>
-  <div>
-    <MaterialNavigation />
-    <MaterialDepartementList />
-  </div>
-</template>
-
-<script>
-import MaterialNavigation from "@/components/material/MaterialNavigation";
-import MaterialDepartementList from "@/components/material/MaterialDepartementList";
-
-export default {
-  name: "MaterialOrder",
-  components: {
-    MaterialNavigation,
-    MaterialDepartementList,
-  },
-};
-</script>
