+++ src/views/material/MaterialDepartementList.vue
@@ -4,61 +4,63 @@
 }
 </style>
 <template>
-  <v-main>
-    <v-card>
-      <v-card-title>Ressort</v-card-title>
-      <v-card-text>
-        <v-select
-          :items="departements"
-          label="Ressort"
-          item-text="name"
-          item-value="id"
-          @change="handleChange"
-        ></v-select>
-      </v-card-text>
-      <v-data-table
-        dense
-        :headers="headers"
-        :items="clients"
-        :items-per-page="20"
-        :footer-props="{
-          'items-per-page-options': [20, 50, -1],
-          showFirstLastPage: true,
-        }"
-        id="clientList"
-        class="elevation-1"
-        @click:row="handleClick"
-      >
-        <template v-slot:item.amount="{ item }">
-          <span>{{
-            item.amount.toLocaleString(undefined, {
-              style: "currency",
-              currency: "CHF",
-            })
-          }}</span>
-        </template>
-      </v-data-table>
-      <v-card-text v-if="showTotal">
-        <span v-if="showClient">{{ total.client }} Kunden, </span
-        >{{ total.order }} Bestellungen, {{ total.orderItem }} Positionen, Summe
-        {{
-          total.amount.toLocaleString(undefined, {
+  <v-card>
+    <portal to="topnav-title">Material / Ressorts</portal>
+    <MaterialNavigation></MaterialNavigation>
+
+    <v-card-title>Ressort</v-card-title>
+    <v-card-text>
+      <v-select
+        :items="departements"
+        label="Ressort"
+        item-text="name"
+        item-value="id"
+        @change="handleChange"
+      ></v-select>
+    </v-card-text>
+    <v-data-table
+      dense
+      :headers="headers"
+      :items="clients"
+      :items-per-page="20"
+      :footer-props="{
+        'items-per-page-options': [20, 50, -1],
+        showFirstLastPage: true,
+      }"
+      id="clientList"
+      class="elevation-1"
+      @click:row="handleClick"
+    >
+      <template v-slot:item.amount="{ item }">
+        <span>{{
+          item.amount.toLocaleString(undefined, {
             style: "currency",
             currency: "CHF",
           })
-        }}
-      </v-card-text>
-    </v-card>
-  </v-main>
+        }}</span>
+      </template>
+    </v-data-table>
+    <v-card-text v-if="showTotal">
+      <span v-if="showClient">{{ total.client }} Kunden, </span
+      >{{ total.order }} Bestellungen, {{ total.orderItem }} Positionen, Summe
+      {{
+        total.amount.toLocaleString(undefined, {
+          style: "currency",
+          currency: "CHF",
+        })
+      }}
+    </v-card-text>
+  </v-card>
 </template>
 
 <script>
 import { apiAuthenticated, ApiError, filter, limit } from "@/lib/api";
 import { joinInPlace } from "@/lib/join";
+import MaterialNavigation from "@/components/material/MaterialNavigation";
 
 export default {
   name: "MaterialDepartementList",
-  components: {},
+  components: { MaterialNavigation },
   data: () => ({
     departements: [],
     headers: [
@@ -74,8 +76,9 @@ export default {
   }),
   methods: {
     handleClick(item) {
+      // TODO replace this with store or URL component
       localStorage.orderSearch = item.client;
-      this.$router.push({ path: "/material/order" });
+      this.$router.push({ name: 'materialOrderList' });
     },
     async handleChange(value) {
       try {
@@ -135,11 +138,7 @@ export default {
           },
           { client: 0, order: 0, orderItem: 0, amount: 0 }
         );
-        if (this.total.client > 1) {
-          this.showClient = true;
-        } else {
-          this.showClient = false;
-        }
+        this.showClient = this.total.client > 1;
         this.showTotal = true;
       } catch (err) {
         if (err instanceof ApiError) {
