--- src/views/material/MaterialSponsoring.vue
@@ -1,19 +1,0 @@
-<template>
-  <div>
-    <MaterialNavigation />
-    <MaterialSponsoringList />
-  </div>
-</template>
-
-<script>
-import MaterialNavigation from "@/components/material/MaterialNavigation";
-import MaterialSponsoringList from "@/components/material/MaterialSponsoringList";
-
-export default {
-  name: "MaterialSponsoring",
-  components: {
-    MaterialNavigation,
-    MaterialSponsoringList,
-  },
-};
-</script>
